#550 Moved HISinOne related commands to new package de.uni_jena.thunibib.his.cli
