+++ common/src/main/java/de/uni_jena/thunibib/his/cli/HISinOneCommandConfiguration.java
@@ -1,4 +1,4 @@
-package de.uni_jena.thunibib;
+package de.uni_jena.thunibib.his.cli;
 
 import de.uni_jena.thunibib.his.api.v1.cs.sys.values.SysValue;
 import org.jdom2.filter.Filters;
