+++ common/src/main/java/de/uni_jena/thunibib/his/cli/HISinOneCommands.java
@@ -1,5 +1,6 @@
-package de.uni_jena.thunibib;
+package de.uni_jena.thunibib.his.cli;
 
+import de.uni_jena.thunibib.Utilities;
 import de.uni_jena.thunibib.his.api.client.HISInOneClient;
 import de.uni_jena.thunibib.his.api.client.HISinOneClientFactory;
 import de.uni_jena.thunibib.his.api.v1.cs.sys.values.SysValue;
