+++ common/src/main/resources/config/thunibib-common/mycore.properties
@@ -33,8 +33,8 @@ MCR.EventHandler.MCRObject.071.Class    =
 MCR.EventHandler.MCRObject.100.Class    = org.mycore.solr.index.MCRSolrIndexEventHandler
 
 # Imports
-MCR.CLI.Classes.External.ThUniBibEnrich=de.uni_jena.thunibib.impex.EnrichmentByAffiliationCommands
-MCR.CLI.Classes.External.ThUniBib=de.uni_jena.thunibib.ThUniBibCommands,de.uni_jena.thunibib.HISinOneCommands
+MCR.CLI.Classes.External.ThUniBibEnrich = de.uni_jena.thunibib.impex.EnrichmentByAffiliationCommands
+MCR.CLI.Classes.External.ThUniBib       = de.uni_jena.thunibib.his.cli.HISinOneCommands
 
 # PPN2DBT-ID
 MCR.MODS.EnrichmentResolver.DataSource.PPN2DBT-ID.IdentifierTypes = ppn
